requirements.txt
# requirements.txt - TeleFrame Python Dependencies

# Core dependencies
# pygame>=2.5.0 do not use pip pygame because of a sdl regression bug in < 2.1.2 version of pygame 
pillow>=10.0.0
pydantic>=2.0.0
toml>=0.10.2

# Telegram bot
python-telegram-bot>=20.0

# File validation and security
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == "win32"

# Async file operations
aiofiles>=23.0.0
aiohttp>=3.8.0

# Optional: For better image processing
opencv-python>=4.8.0  # Optional: Advanced image processing

# Optional: For GPIO support on Raspberry Pi
# RPi.GPIO>=0.7.1; sys_platform == "linux"
# gpiozero>=1.6.2; sys_platform == "linux"

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
